Working on issue #81
